Many hurdles before release...
= License Conflicts =
Turns out GPL/EPL don't mix :-( so we can't use an uberjar and release
one, so update the docs appropriately and the license metadata.

= Noisy data model around transactions =
Simplify down to just a :PERFORMED relationship and ditch the
timestamp stuff for now, PaySim only does 30 days at the moment so no
big deal.

= Transactions lack ordering
Add a post-load threading transformation using CYPHER and use the
latest PaySim to get global step counter.

= Performance dropped after PaySim upgrade =
Remove all the synchronization used for logging to stdout, it's just
noise, and it seems to slow things down by ~10% or so. Users should
know it's running based on their CPU fans kicking in :-)